payments/views.py
from rest_framework import viewsets, permissions
from .models import Payment
from .serializers import PaymentSerializer


class PaymentViewSet(viewsets.ModelViewSet):
    """
    List and retrieve payments for the authenticated user.
    Create sets user to request.user.
    Note: No payment gateway integration yet; status is set by the client.
    For production, integrate provider webhooks and set status from callbacks.
    """
    serializer_class = PaymentSerializer
    permission_classes = [permissions.IsAuthenticated]

    def get_queryset(self):
        return Payment.objects.filter(user=self.request.user).order_by('-created_at')

    def perform_create(self, serializer):
        serializer.save(user=self.request.user)
